test: move some functions into another file
