+++ br/pkg/lightning/backend/external/bench_test.go
@@ -19,12 +19,8 @@ import (
 	"flag"
 	"fmt"
 	"io"
-	"math"
-	"math/rand"
 	"os"
-	"runtime"
 	"runtime/pprof"
-	"slices"
 	"sync"
 	"testing"
 	"time"
@@ -34,240 +30,12 @@ import (
 	"github.com/pingcap/tidb/br/pkg/storage"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/util/intest"
-	"github.com/pingcap/tidb/pkg/util/logutil"
 	"github.com/pingcap/tidb/pkg/util/size"
-	"github.com/stretchr/testify/require"
 	"go.uber.org/atomic"
-	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 )
 
 var testingStorageURI = flag.String("testing-storage-uri", "", "the URI of the storage used for testing")
-
-func openTestingStorage(t *testing.T) storage.ExternalStorage {
-	if *testingStorageURI == "" {
-		t.Skip("testingStorageURI is not set")
-	}
-	s, err := storage.NewFromURL(context.Background(), *testingStorageURI)
-	require.NoError(t, err)
-	return s
-}
-
-type kvSource interface {
-	next() (key, value []byte, handle kv.Handle)
-	outputSize() int
-}
-
-type ascendingKeyGenerator struct {
-	keySize         int
-	keyCommonPrefix []byte
-	count           int
-	curKey          []byte
-	keyOutCh        chan []byte
-}
-
-func generateAscendingKey(
-	count int,
-	keySize int,
-	keyCommonPrefix []byte,
-) chan []byte {
-	c := &ascendingKeyGenerator{
-		keySize:         keySize,
-		count:           count,
-		keyCommonPrefix: keyCommonPrefix,
-		keyOutCh:        make(chan []byte, 100),
-	}
-	c.curKey = make([]byte, keySize)
-	copy(c.curKey, keyCommonPrefix)
-	c.run()
-	return c.keyOutCh
-}
-
-func (c *ascendingKeyGenerator) run() {
-	keyCommonPrefixSize := len(c.keyCommonPrefix)
-	incSuffixLen := int(math.Ceil(math.Log2(float64(c.count)) / 8))
-	if c.keySize-keyCommonPrefixSize < incSuffixLen {
-		panic(fmt.Sprintf("key size %d is too small, keyCommonPrefixSize: %d, incSuffixLen: %d",
-			c.keySize, keyCommonPrefixSize, incSuffixLen))
-	}
-
-	go func() {
-		defer close(c.keyOutCh)
-		for i := 0; i < c.count; i++ {
-			// ret to use most left bytes to alternate the key
-			for j := keyCommonPrefixSize + incSuffixLen - 1; j >= keyCommonPrefixSize; j-- {
-				c.curKey[j]++
-				if c.curKey[j] != 0 {
-					break
-				}
-			}
-			c.keyOutCh <- slices.Clone(c.curKey)
-		}
-	}()
-}
-
-type ascendingKeySource struct {
-	valueSize int
-	keys      [][]byte
-	keysIdx   int
-	totalSize int
-}
-
-func newAscendingKeySource(
-	count int,
-	keySize int,
-	valueSize int,
-	keyCommonPrefix []byte,
-) *ascendingKeySource {
-	keyCh := generateAscendingKey(count, keySize, keyCommonPrefix)
-	s := &ascendingKeySource{
-		valueSize: valueSize,
-		keys:      make([][]byte, count),
-	}
-	for i := 0; i < count; i++ {
-		key := <-keyCh
-		s.keys[i] = key
-		s.totalSize += len(key) + valueSize
-	}
-	return s
-}
-
-func (s *ascendingKeySource) next() (key, value []byte, handle kv.Handle) {
-	if s.keysIdx >= len(s.keys) {
-		return nil, nil, nil
-	}
-	key = s.keys[s.keysIdx]
-	s.keysIdx++
-	return key, make([]byte, s.valueSize), nil
-}
-
-func (s *ascendingKeySource) outputSize() int {
-	return s.totalSize
-}
-
-type ascendingKeyAsyncSource struct {
-	valueSize int
-	keyOutCh  chan []byte
-	totalSize int
-}
-
-func newAscendingKeyAsyncSource(
-	count int,
-	keySize int,
-	valueSize int,
-	keyCommonPrefix []byte,
-) *ascendingKeyAsyncSource {
-	s := &ascendingKeyAsyncSource{
-		valueSize: valueSize,
-		keyOutCh:  generateAscendingKey(count, keySize, keyCommonPrefix),
-	}
-	return s
-}
-
-func (s *ascendingKeyAsyncSource) next() (key, value []byte, handle kv.Handle) {
-	key, ok := <-s.keyOutCh
-	if !ok {
-		return nil, nil, nil
-	}
-	s.totalSize += len(key) + s.valueSize
-	return key, make([]byte, s.valueSize), nil
-}
-
-func (s *ascendingKeyAsyncSource) outputSize() int {
-	return s.totalSize
-}
-
-type randomKeyGenerator struct {
-	keySize         int
-	keyCommonPrefix []byte
-	rnd             *rand.Rand
-	count           int
-	curKey          []byte
-	keyOutCh        chan []byte
-}
-
-func generateRandomKey(
-	count int,
-	keySize int,
-	keyCommonPrefix []byte,
-	seed int,
-) chan []byte {
-	c := &randomKeyGenerator{
-		keySize:         keySize,
-		count:           count,
-		keyCommonPrefix: keyCommonPrefix,
-		rnd:             rand.New(rand.NewSource(int64(seed))),
-		keyOutCh:        make(chan []byte, 100),
-	}
-	c.curKey = make([]byte, keySize)
-	copy(c.curKey, keyCommonPrefix)
-	c.run()
-	return c.keyOutCh
-}
-
-func (c *randomKeyGenerator) run() {
-	keyCommonPrefixSize := len(c.keyCommonPrefix)
-	incSuffixLen := int(math.Ceil(math.Log2(float64(c.count)) / 8))
-	randomLen := c.keySize - keyCommonPrefixSize - incSuffixLen
-	if randomLen < 0 {
-		panic(fmt.Sprintf("key size %d is too small, keyCommonPrefixSize: %d, incSuffixLen: %d",
-			c.keySize, keyCommonPrefixSize, incSuffixLen))
-	}
-
-	go func() {
-		defer close(c.keyOutCh)
-		for i := 0; i < c.count; i++ {
-			c.rnd.Read(c.curKey[keyCommonPrefixSize : keyCommonPrefixSize+randomLen])
-			for j := len(c.curKey) - 1; j >= keyCommonPrefixSize+randomLen; j-- {
-				c.curKey[j]++
-				if c.curKey[j] != 0 {
-					break
-				}
-			}
-			c.keyOutCh <- slices.Clone(c.curKey)
-		}
-	}()
-}
-
-type randomKeySource struct {
-	valueSize int
-	keys      [][]byte
-	keysIdx   int
-	totalSize int
-}
-
-func newRandomKeySource(
-	count int,
-	keySize int,
-	valueSize int,
-	keyCommonPrefix []byte,
-	seed int,
-) *randomKeySource {
-	keyCh := generateRandomKey(count, keySize, keyCommonPrefix, seed)
-	s := &randomKeySource{
-		valueSize: valueSize,
-		keys:      make([][]byte, count),
-	}
-	for i := 0; i < count; i++ {
-		key := <-keyCh
-		s.keys[i] = key
-		s.totalSize += len(key) + valueSize
-	}
-	return s
-}
-
-func (s *randomKeySource) next() (key, value []byte, handle kv.Handle) {
-	if s.keysIdx >= len(s.keys) {
-		return nil, nil, nil
-	}
-	key = s.keys[s.keysIdx]
-	s.keysIdx++
-	return key, make([]byte, s.valueSize), nil
-}
-
-func (s *randomKeySource) outputSize() int {
-	return s.totalSize
-}
 
 type writeTestSuite struct {
 	store              storage.ExternalStorage
@@ -368,40 +136,6 @@ func writeExternalOneFile(s *writeTestSuite) {
 	if s.afterWriterClose != nil {
 		s.afterWriterClose()
 	}
-}
-
-func recordHeapForMaxInUse(filename string) (chan struct{}, *sync.WaitGroup) {
-	doneCh := make(chan struct{})
-	var wg sync.WaitGroup
-	wg.Add(1)
-	maxInUse := uint64(0)
-	go func() {
-		defer wg.Done()
-
-		var m runtime.MemStats
-		ticker := time.NewTicker(300 * time.Millisecond)
-		defer ticker.Stop()
-
-		for {
-			select {
-			case <-doneCh:
-				return
-			case <-ticker.C:
-				runtime.ReadMemStats(&m)
-				if m.HeapInuse <= maxInUse {
-					continue
-				}
-				maxInUse = m.HeapInuse
-				file, err := os.Create(filename)
-				intest.AssertNoError(err)
-				err = pprof.WriteHeapProfile(file)
-				intest.AssertNoError(err)
-				err = file.Close()
-				intest.AssertNoError(err)
-			}
-		}
-	}()
-	return doneCh, &wg
 }
 
 // TestCompareWriter should be run like
@@ -592,51 +326,6 @@ func readFileConcurrently(t *testing.T, s *readTestSuite) {
 	)
 }
 
-func createEvenlyDistributedFiles(
-	store storage.ExternalStorage,
-	fileSize, fileCount int,
-	subDir string,
-) (int, kv.Key, kv.Key) {
-	ctx := context.Background()
-
-	cleanOldFiles(ctx, store, "/"+subDir)
-
-	value := make([]byte, 100)
-	kvCnt := 0
-	var minKey, maxKey kv.Key
-	for i := 0; i < fileCount; i++ {
-		builder := NewWriterBuilder().
-			SetBlockSize(10 * 1024 * 1024).
-			SetMemorySizeLimit(uint64(float64(fileSize) * 1.1))
-		writer := builder.Build(
-			store,
-			"/"+subDir,
-			fmt.Sprintf("%d", i),
-		)
-
-		keyIdx := i
-		totalSize := 0
-		for totalSize < fileSize {
-			key := fmt.Sprintf("key_%09d", keyIdx)
-			if len(minKey) == 0 && len(maxKey) == 0 {
-				minKey = []byte(key)
-				maxKey = []byte(key)
-			} else {
-				minKey = BytesMin(minKey, []byte(key))
-				maxKey = BytesMax(maxKey, []byte(key))
-			}
-			err := writer.WriteRow(ctx, []byte(key), value, nil)
-			intest.AssertNoError(err)
-			keyIdx += fileCount
-			totalSize += len(key) + len(value)
-			kvCnt++
-		}
-		err := writer.Close(ctx)
-		intest.AssertNoError(err)
-	}
-	return kvCnt, minKey, maxKey
-}
-
 func readMergeIter(t *testing.T, s *readTestSuite) {
 	ctx := context.Background()
 	files, _, err := GetAllFileNames(ctx, s.store, "/"+s.subDir)
@@ -745,50 +434,6 @@ func TestCompareReaderEvenlyDistributedContent(t *testing.T) {
 	)
 }
 
-func createAscendingFiles(
-	store storage.ExternalStorage,
-	fileSize, fileCount int,
-	subDir string,
-) (int, kv.Key, kv.Key) {
-	ctx := context.Background()
-
-	cleanOldFiles(ctx, store, "/"+subDir)
-
-	keyIdx := 0
-	value := make([]byte, 100)
-	kvCnt := 0
-	var minKey, maxKey kv.Key
-	for i := 0; i < fileCount; i++ {
-		builder := NewWriterBuilder().
-			SetMemorySizeLimit(uint64(float64(fileSize) * 1.1))
-		writer := builder.Build(
-			store,
-			"/"+subDir,
-			fmt.Sprintf("%d", i),
-		)
-
-		totalSize := 0
-		var key string
-		for totalSize < fileSize {
-			key = fmt.Sprintf("key_%09d", keyIdx)
-			if i == 0 && totalSize == 0 {
-				minKey = []byte(key)
-			}
-			err := writer.WriteRow(ctx, []byte(key), value, nil)
-			intest.AssertNoError(err)
-			keyIdx++
-			totalSize += len(key) + len(value)
-			kvCnt++
-		}
-		if i == fileCount-1 {
-			maxKey = []byte(key)
-		}
-		err := writer.Close(ctx)
-		intest.AssertNoError(err)
-	}
-	return kvCnt, minKey, maxKey
-}
-
 var (
 	objectPrefix = flag.String("object-prefix", "ascending", "object prefix")
 	fileSize     = flag.Int("file-size", 50*units.MiB, "file size")
@@ -796,7 +441,6 @@ var (
 	concurrency  = flag.Int("concurrency", 100, "concurrency")
 	memoryLimit  = flag.Int("memory-limit", 64*units.MiB, "memory limit")
 	skipCreate   = flag.Bool("skip-create", false, "skip create files")
-	fileName     = flag.String("file-name", "test", "file name for tests")
 )
 
 func TestReadFileConcurrently(t *testing.T) {
@@ -867,79 +511,6 @@ func testCompareReaderWithContent(
 	}
 
 	fn(t, suite)
-}
-
-const largeAscendingDataPath = "large_ascending_data"
-
-// TestPrepareLargeData will write 1000 * 256MB data to the storage.
-func TestPrepareLargeData(t *testing.T) {
-	store := openTestingStorage(t)
-	ctx := context.Background()
-
-	cleanOldFiles(ctx, store, largeAscendingDataPath)
-
-	fileSize := 256 * 1024 * 1024
-	fileCnt := 1000
-	keySize := 20
-	valueSize := 100
-	concurrency := runtime.NumCPU() / 2
-	filePerConcUpperBound := (fileCnt + concurrency - 1) / concurrency
-
-	size := atomic.NewInt64(0)
-	now := time.Now()
-	wg := sync.WaitGroup{}
-	wg.Add(concurrency)
-
-	for i := 0; i < concurrency; i++ {
-		i := i
-		go func() {
-			defer wg.Done()
-			writer := NewWriterBuilder().
-				SetMemorySizeLimit(uint64(fileSize)).
-				Build(store, largeAscendingDataPath, fmt.Sprintf("%02d", i))
-			endFile := min((i+1)*filePerConcUpperBound, fileCnt)
-			startFile := min(i*filePerConcUpperBound, endFile)
-			if startFile == endFile {
-				return
-			}
-
-			// slightly reduce total size to avoid generate a small file at the end
-			totalSize := fileSize*(endFile-startFile) - 20*1024*1024
-			kvCnt := totalSize / (keySize + valueSize + 16)
-			source := newAscendingKeyAsyncSource(kvCnt, keySize, valueSize, []byte{byte(i)})
-			key, val, _ := source.next()
-			for key != nil {
-				err := writer.WriteRow(ctx, key, val, nil)
-				intest.AssertNoError(err)
-				size.Add(int64(len(key) + len(val)))
-				key, val, _ = source.next()
-			}
-			err := writer.Close(ctx)
-			intest.AssertNoError(err)
-		}()
-	}
-	wg.Wait()
-	elapsed := time.Since(now)
-	t.Logf("write %d bytes in %s, speed: %.2f MB/s",
-		size.Load(), elapsed, float64(size.Load())/elapsed.Seconds()/1024/1024)
-	dataFiles, _, err := GetAllFileNames(ctx, store, largeAscendingDataPath)
-	intest.AssertNoError(err)
-
-	r, err := store.Open(ctx, dataFiles[0], nil)
-	intest.AssertNoError(err)
-	firstFileSize, err := r.GetFileSize()
-	intest.AssertNoError(err)
-	err = r.Close()
-	intest.AssertNoError(err)
-
-	r, err = store.Open(ctx, dataFiles[len(dataFiles)-1], nil)
-	intest.AssertNoError(err)
-	lastFileSize, err := r.GetFileSize()
-	intest.AssertNoError(err)
-	err = r.Close()
-	intest.AssertNoError(err)
-	t.Logf("total %d data files, first file size: %.2f MB, last file size: %.2f MB",
-		len(dataFiles), float64(firstFileSize)/1024/1024, float64(lastFileSize)/1024/1024)
 }
 
 type mergeTestSuite struct {
@@ -1102,21 +673,4 @@ func TestMergeBench(t *testing.T) {
 	testCompareMergeWithContent(t, 8, createEvenlyDistributedFiles, mergeStep)
 	testCompareMergeWithContent(t, 8, createAscendingFiles, newMergeStep)
 	testCompareMergeWithContent(t, 8, createEvenlyDistributedFiles, newMergeStep)
-}
-
-func TestReadStatFile(t *testing.T) {
-	ctx := context.Background()
-	store := openTestingStorage(t)
-	rd, _ := newStatsReader(ctx, store, *fileName, 4096)
-	for {
-		prop, err := rd.nextProp()
-		if err == io.EOF {
-			break
-		}
-		logutil.BgLogger().Info("read one prop",
-			zap.Int("prop len", prop.len()),
-			zap.Int("prop offset", int(prop.offset)),
-			zap.Int("prop size", int(prop.size)),
-			zap.Int("prop keys", int(prop.keys)))
-	}
 }
